+++ src/ImageSprites/Enums/Optimizations.cs
@@ -1,0 +1,9 @@
+namespace ImageSprites
+{
+    public enum Optimizations
+    {
+        None,
+        Lossless,
+        Lossy
+    }
+}