+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - [x] Command for creating .sprite files
 - [x] Command for updating sprites
 - [x] Generate LESS/Sass/Css file
+- [x] Add JSON schema
+- [x] Image Optimizer integration
+- [ ] .sprite parse error handling
 
 Features that have a checkmark are complete and available for
 download in the